requirements.txt
appium-python-client
python-dotenv
langchain
pandas
numpy
seaborn
matplotlib
scikit-learn
langchain
langchain_community
langchain-core
langchain-classic
langchain_ollama
ipykernel
streamlit
tqdm
